Use plain ES and check for document
